+++ vue.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  css: { extract: true },
-  assetsDir: 'assets',
-}